[WIP] split all handlers failing some tests #146
