+++ server/handlers/all-orgs-view.js
@@ -1,0 +1,18 @@
+var Hoek = require('hoek');
+
+module.exports = (request, reply) => {
+  request.redis.LRANGE('organisations', 0, -1, (error, stringifiedOrgs) => {
+    Hoek.assert(!error, 'redis error');
+    var organisations = {
+      allOrganisations: stringifiedOrgs.map(element => JSON.parse(element)),
+      alternate: [{
+        path: '/orgs/add',
+        name: '+'
+      }, {
+        path: '/people',
+        name: 'People'
+      }]
+    };
+    reply.view('organisations/view', organisations);
+  });
+};
