+++ server/handlers/edit-org.js
@@ -1,0 +1,15 @@
+var Hoek = require('hoek');
+
+module.exports = (request, reply) => {
+  var orgId = request.params.id;
+  request.redis.LINDEX('organisations', orgId, (error, stringifiedOrg) => {
+    Hoek.assert(!error, 'redis error');
+    Hoek.assert(stringifiedOrg, 'Organisation does not exist');
+    var oldOrg = JSON.parse(stringifiedOrg);
+    var orgUpdated = Object.assign({}, oldOrg, request.payload);
+    request.redis.LSET('organisations', orgId, JSON.stringify(orgUpdated), (error, response) => {
+      Hoek.assert(!error, 'redis error');
+      reply.redirect(`/orgs/${orgId}`);
+    });
+  });
+};
