+++ server/handlers/serve-file.js
@@ -1,0 +1,1 @@
+module.exports = (request, reply) => reply.file(request.params.path);
