+++ server/handlers/serve-view.js
@@ -1,0 +1,1 @@
+module.exports = (viewName) => (request, reply) => reply.view(viewName);
