+++ server/models/admin-add-org.js
@@ -1,0 +1,7 @@
+var Joi = require('joi');
+
+module.exports = {
+  payload: {
+    name: Joi.string().min(1).required(),
+  }
+};
