+++ server/models/edit-org.js
@@ -1,0 +1,8 @@
+var Joi = require('joi');
+
+module.exports = {
+  payload: {
+    name: Joi.string().min(1).required(),
+    mission_statement: Joi.any().optional()
+  }
+};
