+++ templates/views/organisations/add.html
@@ -1,0 +1,16 @@
+</div>
+<!-- Include Navbar -->
+{{>navbar}}
+
+<!-- Main Content -->
+<div class="container">
+    <h1 class="title">Add Organisation</h1>
+
+<form class="form" method="post">
+  <label class="form__label" for="name">Organisation Name</label>
+  <input class="form__input" type="text" name="name">
+
+  <button class="button">Submit</button>
+</form>
+
+</div>
